refactor: modularize workflow persistence layer
