+++ docs/architecture.md
@@ -11,6 +11,7 @@ Build a "YouTube of web applications" where each application is sourced from a G
 - **Search & Recommendation API**: Indexes metadata, supports tag-based search (`key:value` pairs), and powers autocomplete suggestions.
 - **Frontend Web App**: Provides a search-first experience with keyboard-centric autocomplete, surfaces app cards, and allows launching previews.
 - **Background Workers**: Handle ingestion and build pipelines, periodic repo sync (polling webhooks), tag enrichment, stale build cleanup, and asset auto-materialization. The ingestion worker hydrates metadata before handing off to a dedicated build worker that can run inline (dev) or via BullMQ (prod). A separate asset materializer worker maintains workflow asset graphs, listens to `asset.produced`/`asset.expired` events, and enqueues runs when freshness policies demand updates.
+- **Workflow repositories**: Workflow persistence is split into focused repositories (`definitions`, `assets`, `runs`, `analytics`) under `services/core/src/workflows/repositories`. HTTP handlers use normaliser helpers in `services/core/src/workflows/http` for request parsing, keeping Fastify concerns separated from SQL. See `docs/workflows-module-structure.md` for the full map.
 - **Timestore Service**: DuckDB-backed time series store that partitions datasets by semantic keys, persists manifests in PostgreSQL, and exposes query/maintenance APIs. It reuses the core Postgres instance while maintaining an isolated schema (`timestore`).
 - **Filestore Service**: Transactional filesystem gateway that tracks local and S3-backed trees, persists node metadata in Postgres, executes mutations via orchestrated executors, and emits Redis events so Metastore and Timestore stay aligned without manual file edits.
 - **Service Registry**: Maintains a collection of auxiliary services (kind, base URL, health, capabilities) in PostgreSQL. Services can be registered declaratively via manifest or at runtime through authenticated API calls, and health polling keeps status changes flowing to subscribers.
