+++ services/core/src/routes/workflows.ts
@@ -17,7 +17,6 @@ import {
   listWorkflowAutoRunsForDefinition,
   updateWorkflowDefinition,
   updateWorkflowRun,
-  getJobDefinitionsBySlugs,
   listWorkflowAssetDeclarationsBySlug,
   listLatestWorkflowAssetSnapshots,
   listWorkflowAssetHistory,
@@ -43,14 +42,10 @@ import {
   listWorkflowRunProducedAssets
 } from '../db/index';
 import type {
-  JobDefinitionRecord,
   JsonValue,
   WorkflowAssetDeclaration,
   WorkflowAssetDeclarationRecord,
   WorkflowAssetSnapshotRecord,
-  WorkflowFanOutTemplateDefinition,
-  WorkflowJobStepBundle,
-  WorkflowJobStepDefinition,
   WorkflowStepDefinition
 } from '../db/types';
 import {
@@ -64,10 +59,6 @@ import {
   workflowTriggerSchema,
   jsonValueSchema,
   workflowAssetPartitionParametersSchema,
-  type WorkflowFanOutTemplateInput,
-  type WorkflowStepInput,
-  type WorkflowTriggerInput,
-  type WorkflowAssetDeclarationInput,
   workflowScheduleCreateSchema,
   workflowScheduleUpdateSchema
 } from '../workflows/zodSchemas';
@@ -78,7 +69,20 @@ import {
 } from '../workflows/partitioning';
 import { computeRunKeyColumns, RUN_KEY_MAX_LENGTH } from '../workflows/runKey';
 import { isRunKeyConflict } from '../db/workflows';
-import { parseBundleEntryPoint } from '../jobs/bundleBinding';
+import {
+  buildWorkflowStepMetadata,
+  normalizeWorkflowSteps,
+  normalizeWorkflowTriggers
+} from '../workflows/http/normalizers';
+import {
+  ANALYTICS_BUCKET_OPTIONS,
+  ANALYTICS_ERROR_MESSAGES,
+  ANALYTICS_RANGE_OPTIONS,
+  mapBucketKeyToInterval,
+  mapIntervalToBucketKey,
+  normalizeAnalyticsQuery,
+  workflowAnalyticsQuerySchema
+} from '../workflows/http/analyticsQuery';
 import {
   enqueueWorkflowRun
 } from '../queue';
@@ -131,9 +135,6 @@ const workflowDefinitionListResponse = jsonResponse.bind(
 );
 const workflowDefinitionResponse = jsonResponse.bind(null, 'WorkflowDefinitionResponse');
 
-type WorkflowJobStepInput = Extract<WorkflowStepInput, { jobSlug: string }>;
-type WorkflowJobTemplateInput = Extract<WorkflowFanOutTemplateInput, { jobSlug: string }>;
-type JobDefinitionLookup = Map<string, JobDefinitionRecord>;
 
 type TimelineTriggerSummary = {
   id: string;
@@ -260,595 +261,6 @@ const DEFAULT_TIMELINE_LIMIT = 200;
 
 const coerceTimestamp = (value: string | null | undefined, fallbackIso: string): string =>
   typeof value === 'string' && value.length > 0 ? value : fallbackIso;
-
-function normalizeAssetPartitioning(
-  partitioning: WorkflowAssetDeclarationInput['partitioning']
-): WorkflowAssetDeclaration['partitioning'] | undefined {
-  if (!partitioning) {
-    return undefined;
-  }
-
-  if (partitioning.type === 'static') {
-    const keys = Array.from(
-      new Set(partitioning.keys.map((key) => key.trim()).filter((key) => key.length > 0))
-    );
-    if (keys.length === 0) {
-      return undefined;
-    }
-    return {
-      type: 'static',
-      keys
-    } satisfies WorkflowAssetDeclaration['partitioning'];
-  }
-
-  if (partitioning.type === 'timeWindow') {
-    const timezone = partitioning.timezone?.trim();
-    const format = partitioning.format?.trim();
-    const normalized: WorkflowAssetDeclaration['partitioning'] = {
-      type: 'timeWindow',
-      granularity: partitioning.granularity,
-      timezone: timezone && timezone.length > 0 ? timezone : undefined,
-      format: format && format.length > 0 ? format : undefined,
-      lookbackWindows:
-        typeof partitioning.lookbackWindows === 'number'
-          ? Math.max(1, Math.floor(partitioning.lookbackWindows))
-          : undefined
-    };
-    return normalized;
-  }
-
-  if (partitioning.type === 'dynamic') {
-    const normalized: WorkflowAssetDeclaration['partitioning'] = { type: 'dynamic' };
-    if (typeof partitioning.maxKeys === 'number') {
-      normalized.maxKeys = Math.max(1, Math.floor(partitioning.maxKeys));
-    }
-    if (typeof partitioning.retentionDays === 'number') {
-      normalized.retentionDays = Math.max(1, Math.floor(partitioning.retentionDays));
-    }
-    return normalized;
-  }
-
-  return undefined;
-}
-
-function buildWorkflowStepMetadata(steps: WorkflowStepDefinition[]) {
-  const metadata = new Map<
-    string,
-    {
-      name: string;
-      type: WorkflowStepDefinition['type'];
-    }
-  >();
-
-  for (const step of steps) {
-    metadata.set(step.id, {
-      name: step.name ?? step.id,
-      type: step.type
-    });
-
-    if (step.type === 'fanout') {
-      const template = step.template;
-      metadata.set(template.id, {
-        name: template.name ?? template.id,
-        type: template.type
-      });
-    }
-  }
-
-  return metadata;
-}
-
-function normalizeAssetDeclarations(
-  declarations: WorkflowAssetDeclarationInput[] | undefined
-): WorkflowAssetDeclaration[] | undefined {
-  if (!declarations || declarations.length === 0) {
-    return undefined;
-  }
-
-  const normalized: WorkflowAssetDeclaration[] = [];
-  const seen = new Set<string>();
-
-  for (const declaration of declarations) {
-    const assetId = declaration.assetId.trim();
-    if (!assetId) {
-      continue;
-    }
-    const key = assetId.toLowerCase();
-    if (seen.has(key)) {
-      continue;
-    }
-    seen.add(key);
-
-    const entry: WorkflowAssetDeclaration = {
-      assetId
-    };
-
-    if (declaration.schema) {
-      entry.schema = declaration.schema;
-    }
-
-    if (declaration.freshness) {
-      entry.freshness = declaration.freshness;
-    }
-
-    if (declaration.autoMaterialize) {
-      entry.autoMaterialize = declaration.autoMaterialize;
-    }
-
-    const partitioning = normalizeAssetPartitioning(declaration.partitioning);
-    if (partitioning) {
-      entry.partitioning = partitioning;
-    }
-
-    normalized.push(entry);
-  }
-
-  return normalized.length > 0 ? normalized : undefined;
-}
-
-function normalizeWorkflowDependsOn(dependsOn?: string[]) {
-  if (!dependsOn) {
-    return undefined;
-  }
-  const unique = Array.from(new Set(dependsOn.map((id) => id.trim()).filter(Boolean)));
-  return unique.length > 0 ? unique : undefined;
-}
-
-function collectWorkflowJobSlugs(steps: WorkflowStepInput[]): string[] {
-  const slugs = new Set<string>();
-  for (const step of steps) {
-    if (step.type === 'service') {
-      continue;
-    }
-    if (step.type === 'fanout') {
-      const template = step.template;
-      if (template.type !== 'service' && typeof template.jobSlug === 'string') {
-        const slug = template.jobSlug.trim().toLowerCase();
-        if (slug) {
-          slugs.add(slug);
-        }
-      }
-      continue;
-    }
-    if (typeof step.jobSlug === 'string') {
-      const slug = step.jobSlug.trim().toLowerCase();
-      if (slug) {
-        slugs.add(slug);
-      }
-    }
-  }
-  return Array.from(slugs);
-}
-
-function lookupJobDefinition(
-  jobDefinitions: JobDefinitionLookup,
-  slug: string | undefined
-): JobDefinitionRecord | undefined {
-  if (!slug) {
-    return undefined;
-  }
-  return jobDefinitions.get(slug.trim().toLowerCase());
-}
-
-function normalizeJobBundle(
-  rawBundle: WorkflowJobStepInput['bundle'] | null | undefined,
-  jobDefinition: JobDefinitionRecord | undefined
-): WorkflowJobStepBundle | null | undefined {
-  if (rawBundle === null) {
-    return null;
-  }
-  const parsed = jobDefinition ? parseBundleEntryPoint(jobDefinition.entryPoint) : null;
-
-  if (rawBundle && rawBundle.strategy === 'latest') {
-    const slugFromInput = typeof rawBundle.slug === 'string' ? rawBundle.slug.trim().toLowerCase() : '';
-    const slug = slugFromInput || parsed?.slug || '';
-    if (!slug) {
-      return parsed
-        ? {
-            strategy: 'latest',
-            slug: parsed.slug,
-            version: null,
-            exportName: parsed.exportName ?? null
-          }
-        : undefined;
-    }
-    const exportName = rawBundle.exportName ?? parsed?.exportName ?? null;
-    return {
-      strategy: 'latest',
-      slug,
-      version: null,
-      exportName
-    } satisfies WorkflowJobStepBundle;
-  }
-
-  if (rawBundle && typeof rawBundle.version === 'string' && rawBundle.version.trim().length > 0) {
-    const slugFromInput = typeof rawBundle.slug === 'string' ? rawBundle.slug.trim().toLowerCase() : '';
-    const slug = slugFromInput || parsed?.slug || '';
-    if (!slug) {
-      return parsed
-        ? {
-            strategy: 'pinned',
-            slug: parsed.slug,
-            version: rawBundle.version.trim(),
-            exportName: rawBundle.exportName ?? parsed.exportName ?? null
-          }
-        : undefined;
-    }
-    const exportName = rawBundle.exportName ?? parsed?.exportName ?? null;
-    return {
-      strategy: 'pinned',
-      slug,
-      version: rawBundle.version.trim(),
-      exportName
-    } satisfies WorkflowJobStepBundle;
-  }
-
-  if (parsed) {
-    return {
-      strategy: 'pinned',
-      slug: parsed.slug,
-      version: parsed.version,
-      exportName: parsed.exportName ?? null
-    } satisfies WorkflowJobStepBundle;
-  }
-
-  return undefined;
-}
-
-function normalizeWorkflowJobStep(
-  step: WorkflowJobStepInput,
-  jobDefinitions: JobDefinitionLookup
-): WorkflowJobStepDefinition {
-  const base = {
-    id: step.id,
-    name: step.name,
-    description: step.description ?? null,
-    dependsOn: normalizeWorkflowDependsOn(step.dependsOn)
-  };
-
-  const jobDefinition = lookupJobDefinition(jobDefinitions, step.jobSlug);
-  const bundle = normalizeJobBundle(step.bundle ?? undefined, jobDefinition);
-  const produces = normalizeAssetDeclarations(step.produces);
-  const consumes = normalizeAssetDeclarations(step.consumes);
-
-  const normalized: WorkflowJobStepDefinition = {
-    ...base,
-    type: 'job',
-    jobSlug: step.jobSlug,
-    parameters: step.parameters ?? undefined,
-    timeoutMs: step.timeoutMs ?? null,
-    retryPolicy: step.retryPolicy ?? null,
-    storeResultAs: step.storeResultAs ?? undefined
-  } satisfies WorkflowJobStepDefinition;
-
-  if (bundle !== undefined) {
-    normalized.bundle = bundle;
-  }
-
-  if (produces) {
-    normalized.produces = produces;
-  }
-
-  if (consumes) {
-    normalized.consumes = consumes;
-  }
-
-  return normalized;
-}
-
-function normalizeWorkflowFanOutTemplate(
-  template: WorkflowFanOutTemplateInput,
-  jobDefinitions: JobDefinitionLookup
-) {
-  const base = {
-    id: template.id,
-    name: template.name,
-    description: template.description ?? null,
-    dependsOn: normalizeWorkflowDependsOn(template.dependsOn)
-  };
-
-  if (template.type === 'service') {
-    const produces = normalizeAssetDeclarations(template.produces);
-    const consumes = normalizeAssetDeclarations(template.consumes);
-
-    const definition: WorkflowFanOutTemplateDefinition = {
-      ...base,
-      type: 'service' as const,
-      serviceSlug: template.serviceSlug.trim().toLowerCase(),
-      parameters: template.parameters ?? undefined,
-      timeoutMs: template.timeoutMs ?? null,
-      retryPolicy: template.retryPolicy ?? null,
-      requireHealthy: template.requireHealthy ?? undefined,
-      allowDegraded: template.allowDegraded ?? undefined,
-      captureResponse: template.captureResponse ?? undefined,
-      storeResponseAs: template.storeResponseAs ?? undefined,
-      request: template.request
-    } satisfies WorkflowFanOutTemplateDefinition;
-
-    if (produces) {
-      definition.produces = produces;
-    }
-
-    if (consumes) {
-      definition.consumes = consumes;
-    }
-
-    return definition;
-  }
-
-  const jobTemplate = template as WorkflowJobTemplateInput;
-  const jobDefinition = lookupJobDefinition(jobDefinitions, jobTemplate.jobSlug);
-  const bundle = normalizeJobBundle(jobTemplate.bundle ?? undefined, jobDefinition);
-  const produces = normalizeAssetDeclarations(jobTemplate.produces);
-  const consumes = normalizeAssetDeclarations(jobTemplate.consumes);
-
-  const normalized: WorkflowFanOutTemplateDefinition = {
-    ...base,
-    type: 'job',
-    jobSlug: jobTemplate.jobSlug,
-    parameters: jobTemplate.parameters ?? undefined,
-    timeoutMs: jobTemplate.timeoutMs ?? null,
-    retryPolicy: jobTemplate.retryPolicy ?? null,
-    storeResultAs: jobTemplate.storeResultAs ?? undefined
-  } satisfies WorkflowFanOutTemplateDefinition;
-
-  if (bundle !== undefined) {
-    normalized.bundle = bundle;
-  }
-
-  if (produces) {
-    normalized.produces = produces;
-  }
-
-  if (consumes) {
-    normalized.consumes = consumes;
-  }
-
-  return normalized;
-}
-
-async function normalizeWorkflowSteps(
-  steps: WorkflowStepInput[]
-): Promise<WorkflowStepDefinition[]> {
-  const jobSlugs = collectWorkflowJobSlugs(steps);
-  const jobDefinitions = await getJobDefinitionsBySlugs(jobSlugs);
-
-  return steps.map((step) => {
-    if (step.type === 'fanout') {
-      const produces = normalizeAssetDeclarations(step.produces);
-      const consumes = normalizeAssetDeclarations(step.consumes);
-
-      const definition: WorkflowStepDefinition = {
-        id: step.id,
-        name: step.name,
-        description: step.description ?? null,
-        dependsOn: normalizeWorkflowDependsOn(step.dependsOn),
-        type: 'fanout' as const,
-        collection: step.collection,
-        template: normalizeWorkflowFanOutTemplate(step.template, jobDefinitions),
-        maxItems: step.maxItems ?? null,
-        maxConcurrency: step.maxConcurrency ?? null,
-        storeResultsAs: step.storeResultsAs ?? undefined
-      } satisfies WorkflowStepDefinition;
-
-      if (produces) {
-        definition.produces = produces;
-      }
-
-      if (consumes) {
-        definition.consumes = consumes;
-      }
-
-      return definition;
-    }
-
-    if (step.type === 'service') {
-      const produces = normalizeAssetDeclarations(step.produces);
-      const consumes = normalizeAssetDeclarations(step.consumes);
-
-      const definition: WorkflowStepDefinition = {
-        id: step.id,
-        name: step.name,
-        description: step.description ?? null,
-        dependsOn: normalizeWorkflowDependsOn(step.dependsOn),
-        type: 'service' as const,
-        serviceSlug: step.serviceSlug.trim().toLowerCase(),
-        parameters: step.parameters ?? undefined,
-        timeoutMs: step.timeoutMs ?? null,
-        retryPolicy: step.retryPolicy ?? null,
-        requireHealthy: step.requireHealthy ?? undefined,
-        allowDegraded: step.allowDegraded ?? undefined,
-        captureResponse: step.captureResponse ?? undefined,
-        storeResponseAs: step.storeResponseAs ?? undefined,
-        request: step.request
-      } satisfies WorkflowStepDefinition;
-
-      if (produces) {
-        definition.produces = produces;
-      }
-
-      if (consumes) {
-        definition.consumes = consumes;
-      }
-
-      return definition;
-    }
-
-    return normalizeWorkflowJobStep(step as WorkflowJobStepInput, jobDefinitions);
-  });
-}
-
-function normalizeWorkflowSchedule(schedule?: WorkflowTriggerInput['schedule']) {
-  if (!schedule) {
-    return undefined;
-  }
-
-  const normalized = {
-    cron: schedule.cron.trim(),
-    timezone: schedule.timezone ? schedule.timezone.trim() : null,
-    startWindow: schedule.startWindow ?? null,
-    endWindow: schedule.endWindow ?? null,
-    catchUp: schedule.catchUp ?? false
-  };
-
-  return normalized;
-}
-
-function normalizeWorkflowTriggers(triggers?: WorkflowTriggerInput[]) {
-  if (!triggers) {
-    return undefined;
-  }
-  return triggers.map((trigger) => {
-    const schedule = normalizeWorkflowSchedule(trigger.schedule);
-    const type = trigger.type.trim();
-    const payload: {
-      type: string;
-      options: JsonValue | null;
-      schedule?: typeof schedule;
-    } = {
-      type,
-      options: (trigger.options ?? null) as JsonValue | null
-    };
-
-    if (schedule) {
-      payload.schedule = schedule;
-    }
-
-    return payload;
-  });
-}
-
-const ANALYTICS_RANGE_OPTIONS = ['24h', '7d', '30d'] as const;
-const ANALYTICS_BUCKET_OPTIONS = ['15m', 'hour', 'day'] as const;
-
-type AnalyticsRangeOption = (typeof ANALYTICS_RANGE_OPTIONS)[number];
-type AnalyticsBucketOption = (typeof ANALYTICS_BUCKET_OPTIONS)[number];
-type AnalyticsRangeKey = AnalyticsRangeOption | 'custom';
-
-const ANALYTICS_RANGE_HOURS: Record<AnalyticsRangeOption, number> = {
-  '24h': 24,
-  '7d': 24 * 7,
-  '30d': 24 * 30
-};
-
-const workflowAnalyticsQuerySchema = z
-  .object({
-    from: z.string().optional(),
-    to: z.string().optional(),
-    range: z.enum(ANALYTICS_RANGE_OPTIONS).optional(),
-    bucket: z.enum(ANALYTICS_BUCKET_OPTIONS).optional()
-  })
-  .partial()
-  .strict();
-
-type WorkflowAnalyticsQuery = z.infer<typeof workflowAnalyticsQuerySchema>;
-
-type NormalizedAnalyticsQuery = {
-  rangeKey: AnalyticsRangeKey;
-  bucketKey: AnalyticsBucketOption | null;
-  options: { from: Date; to: Date; bucketInterval?: string };
-};
-
-const ANALYTICS_ERROR_MESSAGES: Record<string, string> = {
-  invalid_from: 'Invalid "from" timestamp',
-  invalid_to: 'Invalid "to" timestamp',
-  invalid_range: 'The "from" timestamp must be before "to"',
-  invalid_bucket: 'Invalid bucket option'
-};
-
-function parseIsoDate(value: string | undefined): Date | null {
-  if (!value) {
-    return null;
-  }
-  const parsed = new Date(value);
-  if (Number.isNaN(parsed.getTime())) {
-    return null;
-  }
-  return parsed;
-}
-
-function mapBucketKeyToInterval(bucketKey: AnalyticsBucketOption | null | undefined):
-  | { key: AnalyticsBucketOption; interval: string }
-  | null {
-  if (!bucketKey) {
-    return null;
-  }
-  switch (bucketKey) {
-    case '15m':
-      return { key: '15m', interval: '15 minutes' };
-    case 'hour':
-      return { key: 'hour', interval: '1 hour' };
-    case 'day':
-      return { key: 'day', interval: '1 day' };
-    default:
-      return null;
-  }
-}
-
-function mapIntervalToBucketKey(interval: string | null | undefined): AnalyticsBucketOption | null {
-  if (!interval) {
-    return null;
-  }
-  switch (interval) {
-    case '15 minutes':
-      return '15m';
-    case '1 hour':
-      return 'hour';
-    case '1 day':
-      return 'day';
-    default:
-      return null;
-  }
-}
-
-function normalizeAnalyticsQuery(
-  query: WorkflowAnalyticsQuery
-): { ok: true; value: NormalizedAnalyticsQuery } | { ok: false; error: string } {
-  const toDate = parseIsoDate(query.to);
-  if (query.to && !toDate) {
-    return { ok: false, error: 'invalid_to' };
-  }
-  const fromDate = parseIsoDate(query.from);
-  if (query.from && !fromDate) {
-    return { ok: false, error: 'invalid_from' };
-  }
-
-  let rangeKey: AnalyticsRangeKey = query.range ?? '7d';
-  let to = toDate ?? new Date();
-  let from = fromDate ?? null;
-
-  if (fromDate || toDate) {
-    rangeKey = query.range ?? 'custom';
-  }
-
-  const effectiveRange: AnalyticsRangeOption =
-    rangeKey === 'custom' ? '7d' : (rangeKey as AnalyticsRangeOption);
-
-  if (!from) {
-    const hours = ANALYTICS_RANGE_HOURS[effectiveRange] ?? ANALYTICS_RANGE_HOURS['7d'];
-    from = new Date(to.getTime() - hours * 60 * 60 * 1000);
-  }
-
-  if (from.getTime() >= to.getTime()) {
-    return { ok: false, error: 'invalid_range' };
-  }
-
-  const bucketConfig = mapBucketKeyToInterval(query.bucket ?? null);
-  if (query.bucket && !bucketConfig) {
-    return { ok: false, error: 'invalid_bucket' };
-  }
-
-  return {
-    ok: true,
-    value: {
-      rangeKey,
-      bucketKey: bucketConfig?.key ?? null,
-      options: bucketConfig
-        ? { from, to, bucketInterval: bucketConfig.interval }
-        : { from, to }
-    }
-  };
-}
 
 const workflowRunRequestSchema = z
   .object({
